bharat-ev-saathi/requirements.txt
# Bharat EV Saathi - Requirements
# Python 3.9+

# Core Data Science & ML
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Deep Learning (Optional for advanced models)
# tensorflow==2.15.0  # Uncomment if needed

# Web Framework
streamlit==1.29.0
streamlit-option-menu==0.3.6

# Gen AI / LLM
google-generativeai==0.3.2  # Google Gemini API
openai==1.6.1  # Optional: OpenAI API

# Data Visualization
altair==5.2.0
pillow==10.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
kagglehub==0.2.5  # Kaggle dataset downloads

# Geospatial (for charging station maps)
folium==0.15.1
geopy==2.4.1

# Additional utilities
joblib==1.3.2  # For model persistence
scipy==1.11.4

# Development & Testing
pytest==7.4.3
jupyter==1.0.0
notebook==7.0.6
